requirements.txt
# This file is used by pip to install required python packages
# Usage: pip install -r requirements.txt

# Flask Framework
Flask==2.2.2

# Flask Packages
Flask-Login==0.6.2
Flask-Migrate==2.7.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
Flask-User==1.0.2.2
email_validator==1.2.1
Flask-Limiter==2.6.2
Flask-Dance[sqla]==6.2.0

# Other important dependencies
werkzeug==2.2.2

# Automated tests
pytest==3.0.5
pytest-cov==2.4.0

